Finalize test migration
