+++ tests/test_interpret_messages.py
@@ -16,8 +16,6 @@ def mtcars():
 
 
 # Use caplog fixture to capture log messages generated by the interpret logger
-
-
 def test_predictions_list(mtcars, caplog):
     model, idata = mtcars
     caplog.set_level("INFO", logger="__bambi_interpret__")
